Rename RequestModels to Queries
