+++ Application/Queries/GetMessages/GetMessageLookUpDto.cs
@@ -2,7 +2,7 @@
 using Application.Models;
 using AutoMapper;
 
-namespace Application.RequestModels.GetMessages
+namespace Application.Queries.GetMessages
 {
     public record GetMessageLookUpDto : IMapWith<Message>
     {
@@ -10,16 +10,12 @@ namespace Application.RequestModels.GetMessages
         public GetMessageUserDto User { get; init; }
         public string Text { get; init; }
         public DateTime SendTime { get; init; }
-        public int ChatId {get; init; }
         public List<GetAttachmentDto> Attachments { get; init; }
         public List<GetReactionDto> Reactions { get; init; }
 
         public void Mapping(Profile profile)
         {
-            profile.CreateMap<Message, GetMessageLookUpDto>()
-                .ForMember(gmlud => gmlud.ChatId, 
-                    opt => opt
-                        .MapFrom(m => m.Chat.Id));
+            profile.CreateMap<Message, GetMessageLookUpDto>();
         }
     }
 }
