+++ Application/Queries/GetMessages/GetMessageUserDto.cs
@@ -2,7 +2,7 @@
 using Application.Models;
 using AutoMapper;
 
-namespace Application.RequestModels.GetMessages
+namespace Application.Queries.GetMessages
 {
     public record GetMessageUserDto : IMapWith<User>
     {
