+++ Application/Queries/GetServers/GetServersRequest.cs
@@ -1,0 +1,9 @@
+﻿using MediatR;
+
+namespace Application.Queries.GetServer
+{
+    public class GetServersRequest : IRequest<List<GetServerLookupDto>>
+    {
+        public int UserId { get; init; }
+    }
+}
