--- Application/RequestModels/GetServers/GetServersRequest.cs
@@ -1,9 +1,0 @@
-﻿using MediatR;
-
-namespace Application.RequestModels.GetServer
-{
-    public class GetServersRequest : IRequest<List<GetServerLookUpDto>>
-    {
-        public int UserId { get; init; }
-    }
-}
